# openapi.yaml
openapi: 3.0.3
info:
  title: SmartDrop API
  description: |
    SmartDrop backend services — price oracle, webhook delivery, health monitoring, and indexing.

    This specification covers all current REST endpoints and planned future endpoints.
    The public HTTP API is versioned by URL path under `/api/v1`; breaking
    changes are introduced in a new versioned path rather than in place.
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: SmartDrop Labs
    url: https://github.com/SmartDropLabs
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://api.smartdrop.app
    description: Production (planned)
security: []
paths:
...
  /api/v1/airdrops:
    post:
      operationId: createAirdrop
      summary: Create an airdrop
      description: Creates a new airdrop with optional initial recipients.
      tags:
        - Airdrops
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAirdropRequest"
      responses:
        "201":
          description: Airdrop created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Airdrop"
        "400":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      operationId: listAirdrops
      summary: List airdrops
      description: Returns a paginated list of airdrops.
      tags:
        - Airdrops
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Paginated list of airdrops
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Airdrop"
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/airdrops/{id}:
    get:
      operationId: getAirdrop
      summary: Get an airdrop
      description: Returns a single airdrop by ID.
      tags:
        - Airdrops
      parameters:
        - $ref: "#/components/parameters/AirdropIdParam"
      responses:
        "200":
          description: Airdrop details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Airdrop"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      operationId: updateAirdrop
      summary: Update an airdrop
      description: Updates an airdrop's name, description, or expiry ledger.
      tags:
        - Airdrops
      parameters:
        - $ref: "#/components/parameters/AirdropIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAirdropRequest"
      responses:
        "200":
          description: Airdrop updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Airdrop"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteAirdrop
      summary: Delete an airdrop
      description: Permanently deletes an airdrop.
      tags:
        - Airdrops
      parameters:
        - $ref: "#/components/parameters/AirdropIdParam"
      responses:
        "200":
          description: Airdrop deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                    enum: [true]
                  id:
                    type: string
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/airdrops/{id}/cancel:
    post:
      operationId: cancelAirdrop
      summary: Cancel an airdrop
      description: Cancels an airdrop, preventing further execution.
      tags:
        - Airdrops
      parameters:
        - $ref: "#/components/parameters/AirdropIdParam"
      responses:
        "200":
          description: Airdrop cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Airdrop"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/airdrops/{id}/recipients:
    post:
      operationId: addAirdropRecipients
      summary: Add recipients to an airdrop
      description: |
        Adds recipients to an airdrop via JSON body or CSV file upload.
        Duplicate addresses within the request are rejected.
      tags:
        - Airdrops
      parameters:
        - $ref: "#/components/parameters/AirdropIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddRecipientsRequest"
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV file with address and amount columns
      responses:
        "201":
          description: Recipients added
          content:
            application/json:
              schema:
                type: object
                properties:
                  added:
                    type: integer
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          description: Duplicate recipient addresses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      operationId: listAirdropRecipients
      summary: List airdrop recipients
      description: Returns a paginated list of recipients for an airdrop.
      tags:
        - Airdrops
      parameters:
        - $ref: "#/components/parameters/AirdropIdParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Paginated list of recipients
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Recipient"
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    BearerAuth:
      type: apiKey
      name: Authorization
      in: header
      description: |
        API key authentication. Provide your API key with the "Bearer " prefix in the
        `Authorization` header. Keys are created via `POST /api/v1/keys` and
        can be scoped to specific resources.

        Example: `Authorization: Bearer key_1234567890abcdef...`
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: 1-indexed page number.
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Results per page (max 100).
    AirdropIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: ^drop_
      description: Airdrop ID
      example: drop_a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6
  schemas:
    StellarAddress:
      type: string
      pattern: ^G[A-Z0-9]{55}$
      description: Stellar public key (ed25519) starting with G
      example: GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335AX2OBFLDTQLNUEHRGPTM6RIA
...
    AssetCode:
      type: string
      minLength: 1
      maxLength: 12
      pattern: ^[A-Z0-9]+$
      description: Stellar asset code (1–12 uppercase alphanumeric)
      example: USDC
...
    PaginationMeta:
      type: object
      description: |
        The canonical pagination envelope every list endpoint in this API
        returns (see the README's "Pagination" section) — `GET
        /api/v1/webhooks`, `/airdrops`, `/airdrops/:id/recipients`,
        `/alerts`, `/airdrops/:id/onchain-recipients`, and
        `/recipients/:address/claims`. `GET /api/v1/webhooks/:id/deliveries`
        is intentionally exempt (limit-only, no page/offset concept).
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 42
        total_pages:
          type: integer
          example: 3
        has_next:
          type: boolean
          example: true
        has_prev:
          type: boolean
          example: false
      required:
        - page
        - limit
        - total
        - total_pages
        - has_next
        - has_prev
...
    ErrorResponse:
      type: object
      description: >-
        Canonical error envelope. Clients should switch on `error.code`,
        which is stable, rather than on `error.message`, which may be
        reworded at any time.
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - PAYLOAD_TOO_LARGE
                - UNSUPPORTED_MEDIA_TYPE
                - RATE_LIMITED
                - INTERNAL_ERROR
                - UPSTREAM_ERROR
                - SERVICE_UNAVAILABLE
                - AIRDROP_NOT_FOUND
                - AIRDROP_NOT_INDEXED
                - RECIPIENT_LIMIT_EXCEEDED
                - CSV_INVALID_ENCODING
                - CSV_MISSING_COLUMNS
                - CSV_MALFORMED
                - CSV_EMPTY
                - WEBHOOK_NOT_FOUND
                - WEBHOOK_LIMIT_EXCEEDED
                - ALERT_NOT_FOUND
                - API_KEY_NOT_FOUND
                - PRICE_UNAVAILABLE
                - INDEXER_UNAVAILABLE
              example: WEBHOOK_NOT_FOUND
            message:
              type: string
              description: Human-readable error description
              example: Webhook not found
            request_id:
              type: string
              description: Correlates this response with server logs
              example: req_V1StGXR8Z5jdHi6BmyT
            details:
              type: object
              additionalProperties: true
              description: Optional structured context specific to the error code
          required:
            - code
            - message
      required:
        - error
...
    ValidationErrorDetail:
      type: object
      properties:
        error:
          type: string
          example: Validation error
        message:
          type: string
          example: url must be a valid HTTP or HTTPS URL
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              issue:
                type: string
      required:
        - error
        - message
    Airdrop:
      type: object
      properties:
        id:
          type: string
          description: Unique airdrop identifier (prefixed with `drop_`)
          example: drop_a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6
        name:
          type: string
          description: Human-readable airdrop name
          example: Community rewards
        description:
          type: string
          nullable: true
          description: Optional description
        asset:
          $ref: "#/components/schemas/AssetCode"
        asset_issuer:
          $ref: "#/components/schemas/StellarAddress"
        total_amount:
          type: number
          description: Total airdrop amount in asset units
          example: 10000
        expiry_ledger:
          type: integer
          description: Ledger number after which the airdrop expires
          example: 52489123
        contract_airdrop_id:
          type: string
          nullable: true
          description: On-chain contract airdrop ID, set after execution
        status:
          type: string
          enum:
            - draft
            - completed
            - failed
            - cancelled
            - expired
          description: Current airdrop status
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - asset
        - asset_issuer
        - total_amount
        - expiry_ledger
        - status
        - created_at
        - updated_at
    CreateAirdropRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          description: Human-readable airdrop name
          example: Community rewards
        description:
          type: string
          description: Optional description
        asset:
          $ref: "#/components/schemas/AssetCode"
        asset_issuer:
          $ref: "#/components/schemas/StellarAddress"
        total_amount:
          type: number
          minimum: 0
          exclusiveMinimum: true
          description: Total airdrop amount in asset units
          example: 10000
        expiry_ledger:
          type: integer
          description: Ledger number after which the airdrop expires
          example: 52489123
        recipients:
          type: array
          description: Optional initial recipients
          items:
            $ref: "#/components/schemas/Recipient"
      required:
        - name
        - asset
        - asset_issuer
        - total_amount
        - expiry_ledger
    UpdateAirdropRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          description: New airdrop name
        description:
          type: string
          description: New description
        expiry_ledger:
          type: integer
          description: New expiry ledger (must be greater than current ledger)
      minProperties: 1
    AddRecipientsRequest:
      type: object
      properties:
        recipients:
          type: array
          description: Recipients to add to the airdrop
          items:
            $ref: "#/components/schemas/Recipient"
      required:
        - recipients
    Recipient:
      type: object
      properties:
        address:
          $ref: "#/components/schemas/StellarAddress"
        amount:
          type: number
          minimum: 0
          exclusiveMinimum: true
          description: Amount in asset units (at most 7 decimal places)
          example: 100.5
      required:
        - address
        - amount
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorDetail"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: WEBHOOK_NOT_FOUND
              message: Webhook endpoint not found
              request_id: req_V1StGXR8Z5jdHi6BmyT
...
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: RATE_LIMITED
              message: Rate limit exceeded. Try again in 42 seconds.
              request_id: req_V1StGXR8Z5jdHi6BmyT
              details:
                retry_after_seconds: 42
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Maximum requests per window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when the window resets
        X-RateLimit-Tier:
          schema:
            type: string
          description: >-
            Rate limit tier the limit was applied from. Only present when the
            request was metered per API key rather than per IP.
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: INTERNAL_ERROR
              message: An unexpected error occurred
              request_id: req_V1StGXR8Z5jdHi6BmyT